+++ core/src/commonMain/kotlin/dev/lennartegb/vec2compose/core/CommandParser.kt
@@ -3,6 +3,7 @@ package dev.lennartegb.vec2compose.core
 import dev.lennartegb.vec2compose.core.commands.ArcTo
 import dev.lennartegb.vec2compose.core.commands.Close
 import dev.lennartegb.vec2compose.core.commands.Command
+import dev.lennartegb.vec2compose.core.commands.MoveTo
 
 private val regex = "[+-]?\\d*[.]?\\d+".toRegex()
 private val validCommands: CharArray = charArrayOf(
@@ -125,7 +126,7 @@ private fun createMoves(eventString: String, isAbsolute: Boolean): List<Command>
         .validate(2, "Move")
         .windowed(size = 2, step = 2, partialWindows = false)
     val moveTo = commands.first().let {
-        Command.MoveTo(x = it[0], y = it[1], isAbsolute = isAbsolute)
+        MoveTo(x = it[0], y = it[1], isAbsolute = isAbsolute)
     }
     val lineTo = commands.drop(1).map {
         Command.LineTo(
