+++ core/src/commonMain/kotlin/dev/lennartegb/vec2compose/core/commands/MoveTo.kt
@@ -1,0 +1,6 @@
+package dev.lennartegb.vec2compose.core.commands
+
+/**
+ * Move to command indicated by m/M.
+ */
+data class MoveTo(val x: Float, val y: Float, override val isAbsolute: Boolean) : Command
