+++ core/src/commonTest/kotlin/dev/lennartegb/vec2compose/core/imagevector/ComposeMethodCreatorTest.kt
@@ -5,6 +5,7 @@ import dev.lennartegb.vec2compose.core.Scale
 import dev.lennartegb.vec2compose.core.Translation
 import dev.lennartegb.vec2compose.core.commands.Close
 import dev.lennartegb.vec2compose.core.commands.Command
+import dev.lennartegb.vec2compose.core.commands.MoveTo
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
@@ -50,7 +51,7 @@ internal class ComposeMethodCreatorTest {
                     alpha = 0xff
                 ),
                 commands = listOf(
-                    Command.MoveTo(2f, 2f, isAbsolute = false),
+                    MoveTo(2f, 2f, isAbsolute = false),
                     Command.LineTo(4f, 4f, isAbsolute = false),
                     Close
                 ),
@@ -101,7 +102,7 @@ internal class ComposeMethodCreatorTest {
                     alpha = 0xff
                 ),
                 commands = listOf(
-                    Command.MoveTo(2f, 2f, isAbsolute = true),
+                    MoveTo(2f, 2f, isAbsolute = true),
                     Command.LineTo(4f, 4f, isAbsolute = true),
                     Close
                 ),
