+++ core/src/commonTest/kotlin/dev/lennartegb/vec2compose/core/imagevector/ImageVectorParserTest.kt
@@ -3,6 +3,7 @@ package dev.lennartegb.vec2compose.core.imagevector
 import dev.lennartegb.vec2compose.core.ImageVector
 import dev.lennartegb.vec2compose.core.commands.Close
 import dev.lennartegb.vec2compose.core.commands.Command
+import dev.lennartegb.vec2compose.core.commands.MoveTo
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
@@ -21,7 +22,7 @@ internal class ImageVectorParserTest {
                 ImageVector.Path(
                     fillType = ImageVector.Path.FillType.NonZero,
                     commands = listOf(
-                        Command.MoveTo(x = 1f, y = 2f, isAbsolute = true),
+                        MoveTo(x = 1f, y = 2f, isAbsolute = true),
                         Command.LineTo(x = 2f, y = 2f, isAbsolute = true),
                         Command.LineTo(x = 1f, y = 2f, isAbsolute = false),
                         Close
@@ -72,7 +73,7 @@ internal class ImageVectorParserTest {
                 ImageVector.Path(
                     fillType = ImageVector.Path.FillType.EvenOdd,
                     commands = listOf(
-                        Command.MoveTo(x = 1f, y = 2f, isAbsolute = true),
+                        MoveTo(x = 1f, y = 2f, isAbsolute = true),
                         Command.LineTo(x = 2f, y = 2f, isAbsolute = true),
                         Command.LineTo(x = 1f, y = 2f, isAbsolute = false),
                         Close
